+++ tests/test_two_periods_multiple_exog.py
@@ -19,7 +19,16 @@ from toy_models.consumption_retirement_model.final_period_solution import (
 from toy_models.consumption_retirement_model.state_space_objects import (
     get_state_specific_feasible_choice_set,
 )
-from toy_models.consumption_retirement_model.state_space_objects import update_state
+
+from tests.two_period_models.ltc_and_job_offer.dcegm_code import (
+    budget_dcegm_two_exog_processes,
+)
+from tests.two_period_models.ltc_and_job_offer.dcegm_code import func_exog_job_offer
+from tests.two_period_models.ltc_and_job_offer.dcegm_code import func_exog_ltc
+from tests.two_period_models.ltc_and_job_offer.dcegm_code import marginal_utility
+from tests.two_period_models.ltc_and_job_offer.eueler_equation_code import (
+    euler_rhs_two_exog_processes,
+)
 
 WEALTH_GRID_POINTS = 10
 
@@ -28,182 +37,8 @@ TEST_CASES_TWO_EXOG_PROCESSES = list(
 )
 
 
-def flow_util(consumption, choice, params):
-    return consumption ** (1 - params["rho"]) / (1 - params["rho"]) - params[
-        "delta"
-    ] * (1 - choice)
-
-
-def marginal_utility(consumption, params):
-    return consumption ** (-params["rho"])
-
-
-def inverse_marginal_utility(marginal_utility, params):
-    return marginal_utility ** (-1 / params["rho"])
-
-
-def euler_rhs_two_exog_processes(
-    init_cond, params, draws, weights, retirement_choice_1, consumption
-):
-    beta = params["beta"]
-    interest_factor = 1 + params["interest_rate"]
-
-    rhs = 0
-    for index_draw, draw in enumerate(draws):
-        marg_util_draw = marginal_utility_weighted_two_exog_processes(
-            init_cond, params, retirement_choice_1, draw, consumption
-        )
-        rhs += weights[index_draw] * marg_util_draw
-    return rhs * beta * interest_factor
-
-
-def budget_dcegm_two_exog_processes(
-    lagged_choice,
-    ltc,
-    savings_end_of_previous_period,
-    income_shock_previous_period,
-    options,
-    params,
-):
-    # lagged_job_offer = jnp.abs(state[-1] - 2) * (state[-1] > 0) * state[0]  # [1, 3]
-    ltc_patient = ltc == 1  # [2, 3]
-
-    resource = (
-        (1 + params["interest_rate"]) * savings_end_of_previous_period
-        + (params["wage_avg"] + income_shock_previous_period)
-        * (1 - lagged_choice)  # if worked last period
-        - ltc_patient * params["ltc_cost"]
-    )
-    return jnp.maximum(resource, 0.5)
-
-
-def func_exog_ltc(
-    period,
-    ltc,
-    params,
-):
-    prob_ltc = (ltc == 0) * (
-        params["ltc_prob_constant"] + period * params["ltc_prob_age"]
-    ) + (ltc == 1)
-    prob_no_ltc = 1 - prob_ltc
-
-    return jnp.array([prob_no_ltc, prob_ltc])
-
-
-def func_exog_job_offer(
-    job_offer,
-    params,
-):
-    prob_job_offer = (job_offer == 0) * params["job_offer_constant"] + (
-        job_offer == 1
-    ) * (params["job_offer_constant"] + params["job_offer_type_two"])
-    prob_no_job_offer = 1 - prob_job_offer
-
-    return jnp.array([prob_no_job_offer, prob_job_offer])
-
-
-def budget_two_exog_processes(
-    lagged_resources,
-    lagged_consumption,
-    lagged_retirement_choice,
-    wage,
-    bad_health,
-    lagged_job_offer,
-    params,
-):
-    interest_factor = 1 + params["interest_rate"]
-    health_costs = params["ltc_cost"]
-    resources = (
-        interest_factor * (lagged_resources - lagged_consumption)
-        + wage * lagged_job_offer * (1 - lagged_retirement_choice)
-        - bad_health * health_costs
-    ).clip(min=0.5)
-    return resources
-
-
-def choice_prob_retirement(consumption, choice, params):
-    v = flow_util(consumption=consumption, params=params, choice=choice)
-    v_0 = flow_util(consumption=consumption, params=params, choice=0)
-    v_1 = flow_util(consumption=consumption, params=params, choice=1)
-    choice_prob = np.exp(v) / (np.exp(v_0) + np.exp(v_1))
-    return choice_prob
-
-
-def wage(nu, params):
-    wage = params["wage_avg"] + nu
-    return wage
-
-
-def prob_long_term_care_patient(params, lagged_bad_health, bad_health):
-    p = params["ltc_prob"]
-    # jnp.array([[0.7, 0.3], [0, 1]])
-
-    if lagged_bad_health == bad_health == 0:
-        pi = 1 - p
-    elif (lagged_bad_health == 0) and (bad_health == 1):
-        pi = p
-    elif lagged_bad_health == 1 and bad_health == 0:
-        pi = 0
-    elif lagged_bad_health == bad_health == 1:
-        pi = 1
-
-    return pi
-
-
-def prob_job_offer(params, lagged_job_offer, job_offer):
-    # p = params["job_offer_prob"]
-
-    if (lagged_job_offer == 0) and (job_offer == 1):
-        pi = 0.5
-    elif lagged_job_offer == job_offer == 0:
-        pi = 0.5
-    elif lagged_job_offer == 1 and job_offer == 0:
-        pi = 0.1
-    elif lagged_job_offer == job_offer == 1:
-        pi = 0.9
-
-    return pi
-
-
-def marginal_utility_weighted_two_exog_processes(
-    init_cond, params, retirement_choice_1, nu, consumption
-):
-    """Return the expected marginal utility for one realization of the wage shock."""
-    budget_1 = init_cond["wealth"]
-    ltc_state_1 = init_cond["bad_health"]
-    job_state_1 = init_cond["job_offer"]
-
-    weighted_marginal = 0
-    for ltc_state_2 in (0, 1):
-        for job_state_2 in (0, 1):
-            for retirement_choice_2 in (0, 1):
-                budget_2 = budget_two_exog_processes(
-                    budget_1,
-                    consumption,
-                    retirement_choice_1,
-                    wage(nu, params),
-                    ltc_state_2,
-                    job_state_1,
-                    params,
-                )
-
-                marginal_util = marginal_utility(consumption=budget_2, params=params)
-                choice_prob = choice_prob_retirement(
-                    consumption=budget_2, choice=retirement_choice_2, params=params
-                )
-
-                ltc_prob = prob_long_term_care_patient(params, ltc_state_1, ltc_state_2)
-                job_offer_prob = prob_job_offer(params, job_state_1, job_state_2)
-
-                weighted_marginal += (
-                    choice_prob * ltc_prob * job_offer_prob * marginal_util
-                )
-
-    return weighted_marginal
-
-
 @pytest.fixture(scope="module")
-def input_data_two_exog_processes():
+def input_data_two_exog_processes(state_space_functions, utility_functions):
     index = pd.MultiIndex.from_tuples(
         [("utility_function", "rho"), ("utility_function", "delta")],
         names=["category", "name"],
@@ -237,22 +72,12 @@ def input_data_two_exog_processes():
             "choices": np.arange(2),
             "endogenous_states": {
                 "married": [0, 1],
-                # "old": [2, 3],
             },
             "exogenous_processes": {
                 "ltc": {"transition": func_exog_ltc, "states": [0, 1]},
                 "job_offer": {"transition": func_exog_job_offer, "states": [0, 1]},
             },
         },
-    }
-    state_space_functions = {
-        "get_state_specific_choice_set": get_state_specific_feasible_choice_set,
-        "update_endog_state_by_state_and_choice": update_state,
-    }
-    utility_functions = {
-        "utility": flow_util,
-        "inverse_marginal_utility": inverse_marginal_utility,
-        "marginal_utility": marginal_utility,
     }
 
     exog_savings_grid = jnp.linspace(
